# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * ReceiptBud API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface ReceiptItem {
  name: string;
  price: number;
  /** @nullable */
  quantity?: number | null;
  /** @nullable */
  category?: string | null;
}

export interface Receipt {
  id: number;
  userId: string;
  storeName: string;
  date: string;
  total: number;
  /** @nullable */
  tax?: number | null;
  /** @nullable */
  discount?: number | null;
  /** @nullable */
  paymentMethod?: string | null;
  category: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  imageBase64?: string | null;
  isFavorite: boolean;
  items: ReceiptItem[];
  /** @nullable */
  aiInsight?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ReceiptInput {
  storeName: string;
  date: string;
  total: number;
  tax?: number;
  discount?: number;
  paymentMethod?: string;
  category: string;
  notes?: string;
  imageBase64?: string;
  items: ReceiptItem[];
  aiInsight?: string;
}

export interface ReceiptUpdate {
  storeName?: string;
  date?: string;
  total?: number;
  tax?: number;
  discount?: number;
  paymentMethod?: string;
  category?: string;
  notes?: string;
  items?: ReceiptItem[];
}

export interface ReceiptScanInput {
  imageBase64: string;
  mimeType?: string;
}

export interface ReceiptScanResult {
  storeName: string;
  date: string;
  total: number;
  /** @nullable */
  tax?: number | null;
  /** @nullable */
  discount?: number | null;
  /** @nullable */
  paymentMethod?: string | null;
  category: string;
  items: ReceiptItem[];
  /** @nullable */
  aiInsight?: string | null;
}

export type BudgetPeriod = typeof BudgetPeriod[keyof typeof BudgetPeriod];


export const BudgetPeriod = {
  week: 'week',
  month: 'month',
  year: 'year',
} as const;

export interface Budget {
  id: number;
  userId: string;
  category: string;
  limitAmount: number;
  period: BudgetPeriod;
  spent: number;
  createdAt: string;
}

export type BudgetInputPeriod = typeof BudgetInputPeriod[keyof typeof BudgetInputPeriod];


export const BudgetInputPeriod = {
  week: 'week',
  month: 'month',
  year: 'year',
} as const;

export interface BudgetInput {
  category: string;
  limitAmount: number;
  period: BudgetInputPeriod;
}

export type BudgetUpdatePeriod = typeof BudgetUpdatePeriod[keyof typeof BudgetUpdatePeriod];


export const BudgetUpdatePeriod = {
  week: 'week',
  month: 'month',
  year: 'year',
} as const;

export interface BudgetUpdate {
  category?: string;
  limitAmount?: number;
  period?: BudgetUpdatePeriod;
}

export interface DashboardSummary {
  totalSpending: number;
  receiptCount: number;
  largestExpense: number;
  averageReceipt: number;
  topCategory: string;
  budgetRemaining: number;
  savings: number;
  mostVisitedStore: string;
}

export interface CategorySpending {
  category: string;
  amount: number;
  percentage: number;
  count: number;
}

export interface SpendingDataPoint {
  label: string;
  amount: number;
}

export interface Achievement {
  id: string;
  title: string;
  description: string;
  icon: string;
  unlocked: boolean;
  /** @nullable */
  unlockedAt?: string | null;
  xp: number;
  level: number;
}

export interface UserProfile {
  id: string;
  name: string;
  email: string;
  currency: string;
  language: string;
  theme: string;
  notificationsEnabled: boolean;
  totalXp: number;
  level: number;
  createdAt: string;
}

export interface UserProfileUpdate {
  name?: string;
  currency?: string;
  language?: string;
  theme?: string;
  notificationsEnabled?: boolean;
}

export interface GeminiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface GeminiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface GeminiConversationInput {
  title: string;
}

export interface GeminiMessageInput {
  content: string;
}

export interface GeminiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: GeminiMessage[];
}

export interface GeminiImageInput {
  prompt: string;
}

export interface GeminiImageOutput {
  b64_json: string;
  mimeType: string;
}

export interface GeminiError {
  error: string;
}

export type ListReceiptsParams = {
search?: string;
category?: string;
favorite?: boolean;
startDate?: string;
endDate?: string;
sortBy?: string;
sortOrder?: string;
};

export type GetDashboardSummaryParams = {
period?: GetDashboardSummaryPeriod;
};

export type GetDashboardSummaryPeriod = typeof GetDashboardSummaryPeriod[keyof typeof GetDashboardSummaryPeriod];


export const GetDashboardSummaryPeriod = {
  week: 'week',
  month: 'month',
  year: 'year',
} as const;

export type GetSpendingByCategoryParams = {
period?: GetSpendingByCategoryPeriod;
};

export type GetSpendingByCategoryPeriod = typeof GetSpendingByCategoryPeriod[keyof typeof GetSpendingByCategoryPeriod];


export const GetSpendingByCategoryPeriod = {
  week: 'week',
  month: 'month',
  year: 'year',
} as const;

export type GetSpendingOverTimeParams = {
period?: GetSpendingOverTimePeriod;
};

export type GetSpendingOverTimePeriod = typeof GetSpendingOverTimePeriod[keyof typeof GetSpendingOverTimePeriod];


export const GetSpendingOverTimePeriod = {
  week: 'week',
  month: 'month',
  year: 'year',
} as const;