# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ReceiptBud API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /budgets:
    get:
      operationId: listBudgets
      tags: [budgets]
      summary: List all budgets
      responses:
        "200":
          description: List of budgets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Budget"
    post:
      operationId: createBudget
      tags: [budgets]
      summary: Create a budget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetInput"
      responses:
        "201":
          description: Created budget
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
  /budgets/{id}:
    patch:
      operationId: updateBudget
      tags: [budgets]
      summary: Update a budget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetUpdate"
      responses:
        "200":
          description: Updated budget
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteBudget
      tags: [budgets]
      summary: Delete a budget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Dashboard ─────────────────────────────────────────────────────────────
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get spending summary stats
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [week, month, year]
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/spending-by-category:
    get:
      operationId: getSpendingByCategory
      tags: [dashboard]
      summary: Get spending breakdown by category
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [week, month, year]
      responses:
        "200":
          description: Category breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategorySpending"
  /dashboard/spending-over-time:
    get:
      operationId: getSpendingOverTime
      tags: [dashboard]
      summary: Get spending trend over time
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [week, month, year]
      responses:
        "200":
          description: Spending trend data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SpendingDataPoint"

  # ── Achievements ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Receipt schemas ──────────────────────────────────────────────────────
...
    Budget:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        category:
          type: string
        limitAmount:
          type: number
        period:
          type: string
          enum: [week, month, year]
        spent:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - category
        - limitAmount
        - period
        - spent
        - createdAt
    BudgetInput:
      type: object
      properties:
        category:
          type: string
        limitAmount:
          type: number
        period:
          type: string
          enum: [week, month, year]
      required:
        - category
        - limitAmount
        - period
    BudgetUpdate:
      type: object
      properties:
        category:
          type: string
        limitAmount:
          type: number
        period:
          type: string
          enum: [week, month, year]

    # ── Dashboard schemas ────────────────────────────────────────────────────
    DashboardSummary:
      type: object
      properties:
        totalSpending:
          type: number
        receiptCount:
          type: integer
        largestExpense:
          type: number
        averageReceipt:
          type: number
        topCategory:
          type: string
        budgetRemaining:
          type: number
        savings:
          type: number
        mostVisitedStore:
          type: string
      required:
        - totalSpending
        - receiptCount
        - largestExpense
        - averageReceipt
        - topCategory
        - budgetRemaining
        - savings
        - mostVisitedStore
    CategorySpending:
      type: object
      properties:
        category:
          type: string
        amount:
          type: number
        percentage:
          type: number
        count:
          type: integer
      required:
        - category
        - amount
        - percentage
        - count
    SpendingDataPoint:
      type: object
      properties:
        label:
          type: string
        amount:
          type: number
      required:
        - label
        - amount

    # ── Achievement schemas ──────────────────────────────────────────────────
...