# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * ReceiptBud API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalSpending: number;
  receiptCount: number;
  largestExpense: number;
  averageReceipt: number;
  topCategory: string;
  budgetRemaining: number;
  savings: number;
  mostVisitedStore: string;
}